+++ docs/history.md
@@ -7,6 +7,8 @@
 
 - Fireplaces will now be automatically removed on blindfolded babies.
 - Fixed the bug where movable TNT would get duplicated on blindfolded babies. (Thanks sisuka)
+- Fixed bugs with the following babies:
+  - Gold Baby (#15) - The gold hearts are now properly removed.
 
 ### _v1.20.0_ - September 18th, 2023